Adding the Create Todo Endpoint
Also adds Create Todo Endpoint Integration tests.

I fixed a bug in the TodoRepository.cs that was causing an error with returning the created Id.

I also fixed a bug in the TodoApiFactory.cs. I had the username and password values swapped.

I also fixed a bug in the TodoValidator.cs. I was not checking if the description already existed correctly.

+++ src/todo-app.core/Repositories/TodoRepository.cs
@@ -27,7 +27,7 @@ public class TodoRepository(IDbConnectionFactory _dbConnectionFactory) : ITodoRe
             commandType: CommandType.StoredProcedure,
             cancellationToken: token));
 
-        var id = parameters.Get<int>("@TodoId");
+        var id = parameters.Get<long>("@TodoId");
         return id;
     }
